+++ packages/cli/src/commands/deploy.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-import { constants, ethers } from "ethers";
+import { ethers } from "ethers";
 import inquirer from "inquirer";
 import { v4 } from "uuid";
 import { Listr, Logger } from "listr2";
@@ -9,7 +9,10 @@ import fs from "fs";
 import openurl from "openurl";
 import ips from "inquirer-prompt-suggest";
 import { Arguments, CommandBuilder } from "yargs";
+import { generateAndDeploy } from "../utils";
 inquirer.registerPrompt("suggest", ips);
+
+// @dev Note: this deployment command is deprecated and will be removed in a future version. Use `mud deploy-contracts` instead.
 
 // Workaround to prevent tsc to transpile dynamic imports with require, which causes an error upstream
 // https://github.com/microsoft/TypeScript/issues/43329#issuecomment-922544562
@@ -31,7 +34,6 @@ interface Options {
   clientUrl?: string;
   netlifySlug?: string;
   netlifyPersonalToken?: string;
-  upgradeSystems?: boolean;
   codespace?: boolean;
   dry?: boolean;
 }
@@ -53,7 +55,6 @@ export const builder: CommandBuilder<Options, Options> = (yargs) =>
     clientUrl: { type: "string" },
     netlifySlug: { type: "string" },
     netlifyPersonalToken: { type: "string" },
-    upgradeSystems: { type: "boolean" },
     codespace: { type: "boolean" },
     dry: { type: "boolean" },
   });
@@ -73,15 +74,6 @@ function isValidHttpUrl(s: string): boolean {
   }
 
   return url.protocol === "http:" || url.protocol === "https:";
-}
-
-function findLog(deployLogLines: string[], log: string): string {
-  for (const logLine of deployLogLines) {
-    if (logLine.includes(log)) {
-      return logLine.split(log)[1].trim();
-    }
-  }
-  throw new Error("Can not find log");
 }
 
 const getDeployInfo: (args: Arguments<Options>) => Promise<Options> = async (args) => {
@@ -114,21 +106,21 @@ const getDeployInfo: (args: Arguments<Options>) => Promise<Options> = async (arg
     reuseComponents: false,
     deployClient: false,
     clientUrl: "http://localhost:3000",
-    upgradeSystems: false,
   };
 
   const { default: fetch } = await importFetch;
 
+  // DISABLED UNTIL NEW REGISTRY IS READY
   // Fetch deployed lattice chains
-  const latticeChains = args.i
-    ? ((await (await fetch("https://registry.lattice.xyz/api?update=true")).json()) as
-        | { specUrl: string }[]
-        | undefined)
-    : [];
+  // const latticeChains = args.i
+  //   ? ((await (await fetch("https://registry.lattice.xyz/api?update=true")).json()) as
+  //       | { specUrl: string }[]
+  //       | undefined)
+  //   : [];
 
-  const chainSpecs = latticeChains?.map((e) => e.specUrl) || [];
-  console.log("Available Lattice chains");
-  console.log(JSON.stringify(latticeChains, null, 2));
+  // const chainSpecs = latticeChains?.map((e) => e.specUrl) || [];
+  // console.log("Available Lattice chains");
+  // console.log(JSON.stringify(latticeChains, null, 2));
 
   const answers: Options =
     args.upgradeSystems && !args.world
@@ -150,7 +142,7 @@ const getDeployInfo: (args: Arguments<Options>) => Promise<Options> = async (arg
             type: "suggest",
             name: "chainSpec",
             message: "Provide a chainSpec.json location (local or remote)",
-            suggestions: chainSpecs,
+            suggestions: ["https://spec.testnet-chain.linfra.xyz/chainSpec.json"],
             when: () => args.chainSpec == null && config.chainSpec == null,
           },
           {
@@ -191,20 +183,6 @@ const getDeployInfo: (args: Arguments<Options>) => Promise<Options> = async (arg
           },
           {
             type: "list",
-            name: "upgradeSystems",
-            message: "Only upgrade systems?",
-            choices: [
-              { name: "Yes", value: true },
-              { name: "No", value: false },
-            ],
-            default: defaultOptions.upgradeSystems,
-            when: (answers) =>
-              (args.world || config.world || answers.world) &&
-              args.upgradeSystems == null &&
-              config.upgradeSystems == null,
-          },
-          {
-            type: "list",
             name: "reuseComponents",
             message: "Reuse existing components?",
             choices: [
@@ -215,7 +193,6 @@ const getDeployInfo: (args: Arguments<Options>) => Promise<Options> = async (arg
             when: (answers) =>
               !answers.upgradeSystems &&
               !args.upgradeSystems &&
-              !config.upgradeSystems &&
               args.reuseComponents == null &&
               config.reuseComponents == null,
           },
@@ -292,7 +269,6 @@ const getDeployInfo: (args: Arguments<Options>) => Promise<Options> = async (arg
     rpc: args.rpc ?? chainSpec?.rpc ?? config.rpc ?? answers.rpc ?? defaultOptions.rpc,
     wsRpc: args.wsRpc ?? chainSpec?.wsRpc ?? config.wsRpc ?? answers.wsRpc ?? defaultOptions.wsRpc,
     world: args.world ?? chainSpec?.world ?? config.world ?? answers.world,
-    upgradeSystems: args.upgradeSystems ?? config.upgradeSystems ?? answers.upgradeSystems,
     reuseComponents:
       args.reuseComponents ?? config.reuseComponents ?? answers.reuseComponents ?? defaultOptions.reuseComponents,
     deployerPrivateKey: args.deployerPrivateKey ?? config.deployerPrivateKey ?? answers.deployerPrivateKey,
@@ -328,33 +304,6 @@ export const deploy = async (options: Options) => {
   let launcherUrl;
   let worldAddress;
 
-  const cmdArgs = options.upgradeSystems
-    ? [
-        "workspace",
-        "contracts",
-        "forge:deploy",
-        ...(options.dry ? [] : ["--broadcast", "--private-keys", wallet.privateKey]),
-        "--sig",
-        "upgradeSystems(address,address)",
-        wallet.address,
-        options.world || constants.AddressZero,
-        "--fork-url",
-        options.rpc!,
-      ]
-    : [
-        "workspace",
-        "contracts",
-        "forge:deploy",
-        ...(options.dry ? [] : ["--broadcast", "--private-keys", wallet.privateKey]),
-        "--sig",
-        "deploy(address,address,bool)",
-        wallet.address,
-        options.world || constants.AddressZero,
-        options.reuseComponents ? "true" : "false",
-        "--fork-url",
-        options.rpc!,
-      ];
-
   try {
     const tasks = new Listr([
       {
@@ -365,13 +314,16 @@ export const deploy = async (options: Options) => {
               {
                 title: "Contracts",
                 task: async (ctx, task) => {
-                  const child = execa("yarn", cmdArgs);
-                  child.stdout?.pipe(task.stdout());
-                  const { stdout } = await child;
-                  const lines = stdout.split("\n");
+                  const result = await generateAndDeploy({
+                    config: "./deploy.json",
+                    rpc: options.rpc!,
+                    worldAddress: options.world,
+                    deployerPrivateKey: wallet.privateKey,
+                  });
 
-                  ctx.worldAddress = worldAddress = findLog(lines, "world: address");
-                  ctx.initialBlockNumber = findLog(lines, "initialBlockNumber: uint256");
+                  ctx.worldAddress = worldAddress = result.deployedWorldAddress;
+                  ctx.initialBlockNumber = result.initialBlockNumber;
+
                   task.output = chalk.yellow(`World deployed at: ${chalk.bgYellow.black(ctx.worldAddress)}`);
                 },
                 options: { bottomBar: 3 },
@@ -453,14 +405,7 @@ export const deploy = async (options: Options) => {
 
                   launcherUrl = `${clientUrl}?chainId=${options.chainId}&worldAddress=${ctx.worldAddress}&rpc=${options.rpc}&wsRpc=${options.wsRpc}&initialBlockNumber=${ctx.initialBlockNumber}&dev=true&snapshot=&stream=&relay=&faucet=`;
 
-                  // Launcher version:
-                  // `https://play.lattice.xyz?worldAddress=${ctx.worldAddress || ""}&client=${
-                  //   clientUrl || ""
-                  // }&rpc=${options.rpc || ""}&wsRpc=${options.wsRpc || ""}&chainId=${options.chainId || ""}&dev=${
-                  //   options.chainId === 31337 || ""
-                  // }&initialBlockNumber=${ctx.initialBlockNumber}`;
-
-                  if (!options.upgradeSystems) openurl.open(launcherUrl);
+                  openurl.open(launcherUrl);
                 },
                 options: { bottomBar: 3 },
               },