+++ packages/ecs-browser/src/StyledComponents.ts
@@ -1,6 +1,9 @@
-import styled from "styled-components";
+import { setup, styled } from "goober";
+import { createElement } from "react";
 
-export const ComponentBrowserInput = styled.input`
+setup(createElement);
+
+export const ComponentBrowserInput = styled("input")`
   background-color: #383c4a;
   color: #8c91a0;
   border: 1px rgba(0, 0, 0, 0.5) solid;
@@ -19,7 +22,7 @@ export const ComponentBrowserInput = styled.input`
   }
 `;
 
-export const ComponentBrowserButton = styled.button<{ active?: boolean }>`
+export const ComponentBrowserButton = styled("button")<{ active?: boolean }>`
   background-color: ${({ active }) => (active ? "#8c91a0" : "#383c4a")};
   color: ${({ active }) => (active ? "#383c4a" : "#8c91a0")};
   border: 1px rgba(0, 0, 0, 0.5) solid;
@@ -33,7 +36,7 @@ export const ComponentBrowserButton = styled.button<{ active?: boolean }>`
   }
 `;
 
-export const ComponentBrowserSelect = styled.select`
+export const ComponentBrowserSelect = styled("select")`
   width: 180px;
   background-color: #383c4a;
   color: #8c91a0;
@@ -42,18 +45,18 @@ export const ComponentBrowserSelect = styled.select`
   padding: 4px;
 `;
 
-export const ValueForm = styled.form`
+export const ValueForm = styled("form")`
   display: flex;
   flex-direction: row;
   justify-content: space-between;
   align-items: center;
 `;
 
-export const ComponentEditorContainer = styled.div`
+export const ComponentEditorContainer = styled("div")`
   margin: 8px auto;
 `;
 
-export const ComponentTitle = styled.div`
+export const ComponentTitle = styled("div")`
   color: white;
   font-weight: bold;
   display: flex;
@@ -62,12 +65,12 @@ export const ComponentTitle = styled.div`
   justify-content: space-between;
 `;
 
-export const EntityEditorContainer = styled.div`
+export const EntityEditorContainer = styled("div")`
   border-bottom: 1px #8c91a0 solid;
   padding: 8px;
 `;
 
-export const BrowserContainer = styled.div`
+export const BrowserContainer = styled("div")`
   overflow: auto;
   background-color: rgba(27, 28, 32, 1);
   color: #8c91a0;
@@ -75,18 +78,38 @@ export const BrowserContainer = styled.div`
   pointer-events: all;
 `;
 
-export const DraggableNumberLabelContainer = styled.label`
+export const DraggableNumberLabelContainer = styled("label")`
   cursor: ew-resize;
   user-select: none;
   color: #8c91a0;
 `;
 
-export const SmallHeadline = styled.p`
+export const SmallHeadline = styled("p")`
   padding: 8px;
   font-size: 14px;
 `;
 
-export const Collapse = styled.div<{ opened: boolean }>`
+export const Collapse = styled("div")<{ opened: boolean }>`
   height: ${({ opened }) => (opened ? "auto" : "0px")};
   overflow: ${({ opened }) => (opened ? "initial" : "hidden")};
+`;
+
+export const SyntaxHighlighterWrapper = styled("div")`
+  .shiki {
+    background-color: #282a36;
+
+    font-family: Consolas, Monaco, "Andale Mono", "Ubuntu Mono", monospace;
+
+    line-height: 1.5;
+    tab-size: 4;
+    hyphens: none;
+    padding: 1em;
+    margin: 0.5em 0px;
+    overflow: auto;
+    border-radius: 0.3em;
+    overflow: auto;
+    overflow-wrap: break-word;
+    white-space: pre-wrap;
+    text-shadow: rgba(0, 0, 0, 0.3) 0px 1px;
+  }
 `;