+++ packages/solecs/src/Indexer.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: Unlicense
+// SPDX-License-Identifier: MIT
 pragma solidity >=0.8.0;
 
 import { IERC165 } from "./interfaces/IERC165.sol";