+++ packages/solecs/src/interfaces/IWorld.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: Unlicense
+// SPDX-License-Identifier: MIT
 pragma solidity >=0.8.0;
 import { QueryType } from "./Query.sol";
 import { IUint256Component } from "./IUint256Component.sol";
@@ -44,4 +44,6 @@ interface IWorld {
   function getUniqueEntityId() external view returns (uint256);
 
   function query(WorldQueryFragment[] calldata worldQueryFragments) external view returns (uint256[] memory);
+
+  function init() external;
 }